requirements.txt
twilio==7.15.4
flask==2.0.1
werkzeug==2.0.1
python-dotenv==1.0.0
pytest==7.3.1
pydantic==1.10.7
loguru==0.7.0
requests==2.28.2 